test: unify provider and selector tests
EX-6762
